pyproject.toml
[tool.poetry]
name = "momok-backend"
version = "0.1.0"
description = ""
authors = ["judlejudle <judean2021@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "momok_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
django-phonenumber-field = "^7.0.0"
django-phonenumbers = "^1.0.1"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
